feat: remove unused modules
