+++ go.mod
@@ -13,8 +13,4 @@ require (
 	github.com/iancoleman/strcase v0.2.0
 	github.com/klauspost/cpuid/v2 v2.2.4
 	github.com/stretchr/testify v1.8.2
-	github.com/thrift-iterator/go v0.0.0-20190402154806-9b5a67519118
-	github.com/v2pro/plz v0.0.0-20221028024117-e5f9aec5b631 // indirect
-	github.com/v2pro/quokka v0.0.0-20171201153428-382cb39c6ee6 // indirect
-	github.com/v2pro/wombat v0.0.0-20180402055224-a56dbdcddef2 // indirect
 )
